+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 * Added `compas_blender.conversions.BlenderMesh`.
 * Added option to return strip faces from `compas.datastructure.Halfedge.edge_strip`.
 * Added `compas.geometry.Bezier.transform`.
+* Added `rgba` and `rgba255` properties to `compas.colors.Color`.
 
 ### Changed